Taken Page_CharacterReceived to the standard layout mode
